requirements.txt
langchain==0.3.8
langchain-core==0.3.21
langchain-community==0.3.8
langchain-text-splitters==0.3.2
langchain-google-genai==2.0.5
google-generativeai==0.8.3
google-ai-generativelanguage==0.6.10
pydantic==2.9.2
pydantic-settings==2.10.1
python-dotenv==1.0.1
httpx-sse==0.4.1
duckduckgo-search
wikipedia


